chore: Fix AppVeyor config + drop support for node 4 (#304)
It seems that the current AppVeyor config is broken because npm received an upgrade which breaks the build (and it shouldn't). The culprit is the line which installs the latest version of npm (while being used e.g. with node 4). It might be worth to make the config future-proof instead.